+++ typings/SocialLink.d.ts
@@ -1,0 +1,5 @@
+export type SocialLink = {
+    id: number,
+    url: string,
+    isActive: boolean
+}